+++ crates/hc/src/setups.rs
